+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/exception/CodedException.kt
@@ -145,9 +145,19 @@ internal class ArgumentCastException(
   providedType: ReadableType,
   cause: CodedException,
 ) : DecoratedException(
-  message = "Argument at index '$argIndex' couldn't be casted to type '$argDesiredType' (received '$providedType').",
+  message = "The ${formatOrdinalNumber(argIndex + 1)} argument cannot be cast to type $argDesiredType (received $providedType)",
   cause,
-)
+) {
+  companion object {
+    fun formatOrdinalNumber(number: Int) = "$number" + when {
+      (number % 100 in 11..13) -> "th"
+      (number % 10) == 1 -> "st"
+      (number % 10) == 2 -> "nd"
+      (number % 10) == 3 -> "rd"
+      else -> "th"
+    }
+  }
+}
 
 internal class FieldCastException(
   fieldName: String,